backend/requirements.txt
fastapi
uvicorn
python-multipart
google-generativeai
pydantic
pymupdf
python-dotenv
pillow
fpdf
